foreign keys modifications

+++ src/main/java/guru/springframework/recipeproject/repositories/BuildingRepository.java
@@ -3,6 +3,8 @@ package guru.springframework.recipeproject.repositories;
 import guru.springframework.recipeproject.models.Building;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BuildingRepository extends CrudRepository<Building, Long> {
+import java.math.BigInteger;
+
+public interface BuildingRepository extends CrudRepository<Building, Integer> {
 
 }
