+++ src/main/java/guru/springframework/recipeproject/repositories/PasswordResets.java
@@ -3,6 +3,8 @@ package guru.springframework.recipeproject.repositories;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PasswordResets extends CrudRepository<guru.springframework.recipeproject.models.PasswordResets, Long> {
+import java.math.BigInteger;
+
+public interface PasswordResets extends CrudRepository<guru.springframework.recipeproject.models.PasswordResets, Integer> {
 
 }
