+++ src/main/java/guru/springframework/recipeproject/repositories/UnitRepository.java
@@ -3,6 +3,8 @@ package guru.springframework.recipeproject.repositories;
 import guru.springframework.recipeproject.models.Unit;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UnitRepository extends CrudRepository<Unit, Long> {
+import java.math.BigInteger;
+
+public interface UnitRepository extends CrudRepository<Unit, Integer> {
 
 }
