+++ src/main/java/guru/springframework/recipeproject/repositories/UsersRepository.java
@@ -3,6 +3,8 @@ package guru.springframework.recipeproject.repositories;
 import guru.springframework.recipeproject.models.Users;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UsersRepository extends CrudRepository<Users, Long> {
+import java.math.BigInteger;
+
+public interface UsersRepository extends CrudRepository<Users, Integer> {
 
 }
